adds BoGL-specific syntax highlighting
